requirements.txt
streamlit==1.31.1
assemblyai==0.17.0
python-docx==1.1.0
reportlab==4.0.9
python-dotenv==1.0.0
transformers==4.37.2 